pyproject.toml
[project]
name = "wfgen"
version = "0.1.0"
description = "beta"
readme = "README.md"
requires-python = ">=3.8"
keywords = []
authors = [
    { name = "B. Clark", email = "bill.clark@vt.edu" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: TBD",
    "Programming Language :: Python :: 3"
]
dependencies = [
    "pyzmq",
    "pyyaml",
    "numpy",
    "scipy",
    "matplotlib",
    "pydub",
    "tqdm",
    "psutil"
]

[project.scripts]
wfgen_server = "wfgen.server:main"
wfgen_cli = "wfgen.client:main"

[tool.setuptools]
script-files = [
    "scripts/consolidate_reports.py",
    "scripts/proc_query.py",
    "scripts/wav_server.py",
    "scripts/wfgen_env.sh"
]
package-dir = {"" = "py_src"}


[build-system]
requires=["setuptools>=61.0.0","wheel","build"]
